requirements.txt
flask
flask-sock
websocket-client
psutil
